Move eventcachemetrics inside pkg/eventcache
There are no functional changes in this commit, only moving code around.
